% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CI.R
\name{add_CI}
\alias{add_CI}
\title{Fill in the confidence interval (CI) column}
\usage{
add_CI(timber)
}
\arguments{
\item{timber}{a tibble of timber.}
}
\value{
A tibble of timber with a complete \emph{oddsci} column.
}
\description{
\code{add_CI} fills in any blank cells in the confidence interval column
  with the assumed value of 95.
}
