% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_HTMLink.R
\name{add_HTMLink}
\alias{add_HTMLink}
\title{Generate HTML Links Using a Document Identifier}
\usage{
add_HTMLink(timber)
}
\arguments{
\item{timber}{a tibble of timber, with a table built by \code{\link{build_table}}.}
}
\value{
The passed tibble of timber with an additional column containing HTML Link(s).
}
\description{
\code{add_HTMLink()} adds a column to the passed tibble, containing the HTML Link
   corresponding to the document identifier provided.
}
\details{
A document identifier is a unique sequence of letters and/or numbers
   assigned to a published work to facilitate its unambiguous reference.
   Generally, these identifiers allow for a given work to be assigned a
   standardized URL. The \code{add_HTMLink} function generates the HTML
   link for the following supported document identifier standards:

   \itemize{
       \item DOI (Digital Object Identifier)
       \item PMID (PubMed ID)
   }

   The \code{add_HTMLink} function works by detecting the appropriate
   document identifier type, using regular expressions. The DOI is identified
   as beginning with two numerical digits, followed by a period, four or five numerical
   digits, and a forward slash. The PMID is a series of eight numerical digits.
   There is no additional validation done; while a non-DOI is unlikely to be
   inadvertently recognized as a DOI, a non-PMID string of eight numbers could
   easily be mistaken.
}
