% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ident.R
\name{add_ident}
\alias{add_ident}
\title{Generate a Suggested Analytica Identifier}
\usage{
add_ident(query, cedar_version)
}
\arguments{
\item{query}{a tibble of timber, with a table built by \code{\link{build_table}}.}

\item{cedar_version}{numeric [1,2]: the version of CEDAR that created the export.}
}
\value{
The passed tibble of timber with an additional column containing the generated
   identifiers.
}
\description{
\code{add_ident()} adds a column to the passes data frame, containing an
   identifier for use in Analytica.
}
\details{
Each node in an Analytica model requires an identifier - a unique string,
   of max 20 characters, with no spaces. \code{add_ident} generates the
   string, which includes the factor (or meta-analysis number), and is based
   on the title of the node.

   The identifier is generated in the format "R00001_name_here".
}
