% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_scraps.R
\name{append_scraps}
\alias{append_scraps}
\title{Append the Scrap Pile to the Processed Timber}
\usage{
append_scraps(timber)
}
\arguments{
\item{timber}{a tibble of timber, returned from \code{\link{add_ident}}.}
}
\value{
The passed tibble of processed timber with additional rows containing the unusable
   factors from the scrap pile.
}
\description{
\code{append_scraps()} adds the unusable factors in the scrap pile to the end of the
  processed timber.
}
