% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_chairs.R
\name{build_chairs}
\alias{build_chairs}
\title{Calculate the Odds Ratio (OR) and the SE(log(OR))}
\usage{
build_chairs(timber, log_base = 10)
}
\arguments{
\item{timber}{a tibble of timber, with a table built by \code{\link{build_table}}.}

\item{log_base}{numeric: the logarithm base used in the calculation of the SE(log(OR)) from CIs.}
}
\value{
A tibble of timber with additional columns: \emph{odds_ratio} and \emph{se_log_or}.
}
\description{
\code{build_chairs()} calculates the odds ratio [OR] and standard error of the log(odds ratio)
  or SE(log(OR)).
}
\details{
The odds ratio and the standard error of the log(odds ratio) can be calculated using a complete
  contingency table, or the odds ratio and the confidence intervals (CIs).

  \subsection{Contingency Table}{
    The formula to calculate the odds ratio from a contingency table is:

    \eqn{OR = (A/B) / (C/D)}

    The formula to calculate the standard error of the log(odds ratio) from a contingency table is:

    \eqn{SE(log(OR)) = sqrt((1/A) + (1/B) + (1/C) + (1/D))}

  }

  \subsection{Odds Ratio and CIs}{
    The formula to calculate the standard error of the log(odds ratio) from the CIs is:

    \eqn{SE(log(OR)) = (log(oddsup) - log(oddslo)) / (2 * Z)}

    \emph{where Z is the Z value corresponding to 1/2 alpha, and alpha = 1 - confidence level}.

    Note, the the choice of logarithm base for this calculation may be specified by the user as
    \code{log_base} (default = 10).

  }
}
