% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_horse.R
\name{build_horse}
\alias{build_horse}
\title{Calculate the significance of a factor}
\usage{
build_horse(timber)
}
\arguments{
\item{timber}{a tibble of timber, with a table built by \code{\link{build_table}}.}
}
\value{
A tibble of timber with additional columns: \emph{pval}.
}
\description{
\code{build_horse()} calculates the significance of the OR of a count or prevalence-based factor, using
  the Fisher's Exact Test. It does not calculate the significance of the OR if the factor has an
  \emph{odds_ratio} grain.
}
