% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_version.R
\name{check_version}
\alias{check_version}
\title{Check that the CEDAR version specified is valid}
\usage{
check_version(cedar_version)
}
\arguments{
\item{cedar_version}{numeric [1, 2]: the CEDAR version.}
}
\description{
\code{check_version} checks if the CEDAR version specified is valid, and
  prompts the user to correct the input.
}
