% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupe_MA.R
\name{dedupe_MA}
\alias{dedupe_MA}
\title{Remove Duplicate Rows for Meta-Analysis}
\usage{
dedupe_MA(query)
}
\arguments{
\item{query}{a tibble of timber, with a table built by \code{\link{build_table}}.}
}
\value{
A tibble of timber, with duplicate rows removed.
}
\description{
\code{dedupe_MA()} removes duplicate rows to retain the largest available
   meta-analysis group. It was adapted from the now depreciated cedarr
   package for Sawmill.
}
\details{
A factor may be included in more than one meta-analysis. For example, a
   factor describing resistance to a third-generation cephalosporin may be
   included in a meta-analysis across other third-generation cephalosporins,
   but may also be included in a larger meta-analysis across all
   cephalosporins. Where a factor is included in multiple meta-analyses
   applicable to a query, duplicate rows will be returned.
   The dedupeMA function removes these duplicate rows, retaining the largest
   meta-analysis group available. If the groups are of the same size, they are
   both retained.
   The dedupeMA function first separates those rows involved in meta-analysis
   from those uninvolved into two tibbles. It then searches for duplicate
   factors. If duplicates exist, the associated meta-analyses are identified
   for each duplicate, and the size of each meta-analysis is determined. Those
   which are not the largest are dropped from the 'involved' tibble, and the
   two tibbles are then reunited.
}
