% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{diam}
\alias{diam}
\title{Calculate Mean Diameters}
\usage{
diam(x, method = c("ar", "qmd", "gfz", "do", "d200", "dw"), s = 1, a = 1)
}
\arguments{
\item{x}{a vector of diameters}

\item{method}{"ar" for arithmetic, "qmd" for quadratic mean
diameter, "gfz" for Grundflaechenzentralstamm, "do" for dominant
diameter based on basal area of 100 largest trees per ha, "d200"
for dominant diameter based on basal area of 200 largest trees per
ha, and "dw" for Weise'sche Oberhoehe.}

\item{s}{class size (integer)}

\item{a}{size of area in ha}
}
\value{
a numeric vector with diameters
}
\description{
Calculate mean diameters in many different ways
}
\details{
Provides some commonly used mean diameters for forest growth and
yield science.
}
\examples{
d <- rnorm(40, mean = 50, sd = 3)
diam(d)
}
