% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{diam_group}
\alias{diam_group}
\title{Group Diameters}
\usage{
diam_group(x, s = 1)
}
\arguments{
\item{x}{numeric vector of diameters}

\item{s}{an integer for specifying group size}
}
\value{
a \code{data.table} with the diameters, the group they
belong to and the respective upper group boundary
}
\description{
Group diameters in diameter classes of given size
}
\details{
This function takes a numeric vector of diameters and a group size
and classifies each diameter into a group.
}
\examples{
d <- rnorm(40, mean = 50, sd = 3)
diam_group(d)
}
