% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ma_results.R
\name{format_ma_results}
\alias{format_ma_results}
\title{Format all meta-analyses results}
\usage{
format_ma_results(ma_results, ma_ids)
}
\arguments{
\item{ma_results}{list: the raw meta-analysis results provided by \code{\link{do_MA}}.}

\item{ma_ids}{a vector containing each unique meta-analysis ID in the timber.}
}
\description{
\code{format_ma_results()} converts the full meta-analysis results (all 100 parameters)
   into a tabular format, and organizes them by meta-analysis ID, so that they are ready to be
   written to a csv file. It then saves them to the global environment.
}
