% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_types.R
\name{get_col_types}
\alias{get_col_types}
\title{Get the Expected Column Types for an Input CEDAR Query}
\usage{
get_col_types(cedar_version)
}
\arguments{
\item{cedar_version}{numeric [1,2]: the version of CEDAR that created the export.}
}
\value{
A vector the size of the number of columns in the export version, indicating
  either "text" or "numeric" for the expected data type of each column
}
\description{
The \code{get_col_types()} function gets the expected data types for each column in the
 CEDAR query version.
}
