% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_col_warnings.R
\name{handle_col_warnings}
\alias{handle_col_warnings}
\title{Handles Warnings Generated by Unexpected Data Types in Input Queries}
\usage{
handle_col_warnings()
}
\description{
The \code{handle_col_warnings()} function gives the user the option to stop
 the processing pipeline in order to fix warnings generated by the input CEDAR query.
 It is called after the \code{\link{log_warnings}} function is used to print warning messages
 that result when data in one or more of the input query's columns does not match the
 expected data type for those column(s) (as dictated by \code{\link{get_col_types}}). These
 messages indicate the exact cell locations responsible, to aid in the user's decision
 to stop or continue processing, as well as the process of fixing the specified cells,
 if applicable.
}
