% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_timber.R
\docType{data}
\name{local_timber}
\alias{local_timber}
\title{A CEDAR timber}
\format{
A data frame with columns as specified in raw_timber_specs.csv.
}
\usage{
local_timber
}
\description{
A cached version of a CEDAR timber with all resistance outcomes. The cached
  version is updated using update_local_timber().
}
\keyword{datasets}
