% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mill.R
\name{mill}
\alias{mill}
\title{The sawmill Pipeline}
\usage{
mill(
  timber,
  cedar_version = 2,
  write_scrap = TRUE,
  low_cell_correction = 0.5,
  low_cell_threshold = 0
)
}
\description{
The \code{mill()} runs the CEDAR pipeline with useful defaults.
}
