% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_fields.R
\name{reorder_fields}
\alias{reorder_fields}
\title{Reorder the processed timber's fields}
\usage{
reorder_fields(timber, cedar_version)
}
\arguments{
\item{timber}{a tibble of timber, returned from \code{\link{add_ident}}.}

\item{cedar_version}{numeric [1,2]: the version of CEDAR that created the export.}
}
\value{
The passed tibble of processed timber with reorganized columns, ready for Analytica.
}
\description{
\code{reorder_fields()} organizes the left-to-right placement of the fields (columns) to match
  the order in which they are arranged in the iAM.AMR Analytica model nodes. It also deletes the
  fields \code{exclude_sawmill} and \code{exclude_sawmill_reason} from the final output, as
  these are only needed internally, to filter out unusable factors into the \code{scrap_pile}.
}
