% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_csv}
\alias{save_csv}
\title{Saves the passed tibble to a csv file}
\usage{
save_csv(result_name, tibble_save)
}
\arguments{
\item{result_name}{a description of the type of data being saved (i.e. the processed timber).}

\item{tibble_save}{the tibble of data to be saved.}
}
\value{
a csv file, saved to a location of the user's choice.
}
\description{
\code{save_csv()} saves the passed tibble of data to a csv file, prompting the
  user for a file name, as well as the file path to a chosen save location along the way.
}
