% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sawmill.R
\docType{package}
\name{sawmill}
\alias{sawmill}
\title{sawmill: A Package to Process CEDAR Queries}
\description{
The \code{sawmill} package is used to process queries exported from CEDAR, herein refered to
  as \strong{timber}.
}
