% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_mill.R
\name{sub_mill}
\alias{sub_mill}
\title{Run a Substep of the sawmill Pipeline, with Error Handling}
\usage{
sub_mill(mill_step, step_name)
}
\arguments{
\item{mill_step}{an expression that calls a function that represents a main step in the
sawmill pipeline}

\item{step_name}{string: the name of the function being called in the \code{mill_step} parameter}
}
\description{
The \code{sub_mill()} runs a substep in the sawmill pipeline, providing a message to
 the user in the event of an error. This message will specify which function caused
 the error.
}
