% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{volume}
\alias{volume}
\title{Stem volume}
\usage{
volume(d, h, species = "Fagus sylvatica")
}
\arguments{
\item{d}{numeric vector with stem diameters in 1.3 m height in cm}

\item{h}{numeric vectors with height in m}

\item{species}{character}
}
\value{
numeric vector with volume of the trees in cubic metres
}
\description{
Calculate stem volume based on height, diameter and species
}
\details{
This function uses the approximation function of Franz et al. (1973) to
calculate the stem volume based on the species specific form number. It is
estimated in two steps: a first level of functions is derived by species and
diameter, then this set of functions enters the final approximation function
which depends on height alone.

At the moment, only beech and spruce are implemented.
}
\examples{
volume(30:35, 20:25, "Fagus sylvatica")
}
