% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CI.R
\name{add_CI}
\alias{add_CI}
\title{Fill in the Confidence Interval (CI) Column}
\usage{
add_CI(timber)
}
\arguments{
\item{timber}{a tibble of timber.}
}
\value{
A tibble of timber with a complete \emph{oddsci} column.
}
\description{
\code{add_CI} fills in empty cells in the confidence interval column with the
  value 95.
}
