% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CI.R
\name{add_CI}
\alias{add_CI}
\title{Add a Confidence Interval (CI) Column to Timber}
\usage{
add_CI(timber)
}
\arguments{
\item{timber}{a tibble of timber.}
}
\value{
A tible of timber with additional column \emph{oddsci}.
}
\description{
\code{add_CI} checks if the CEDAR version specified is 1 or 2, else throws an error.
}
