% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d.R
\name{diam}
\alias{diam}
\title{Calculate Mean Diameters}
\usage{
diam(x, method = c("ar", "qmd", "gfz", "dom"), s = 1, a = 1, n = 100)
}
\arguments{
\item{x}{a vector of diameters}

\item{method}{"ar" for arithmetic, "qmd" for quadratic mean
diameter, "gfz" for Grundflaechenzentralstamm, "dom" for dominant
diameter}

\item{s}{class size (integer)}

\item{a}{size of area in ha}

\item{n}{number of trees/ha for dominant diameter}
}
\value{
a numeric vector with diameters
}
\description{
Calculate mean diameters in many different ways
}
\details{
Provides some commonly used mean diameters for forest growth and
yield science.
}
\examples{
d <- rnorm(40, mean = 50, sd = 3)
diam(d)
}

