% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mill.R
\name{mill}
\alias{mill}
\title{The sawmill Pipeline}
\usage{
mill(
  timber_path,
  cuts,
  col_data_types,
  cedar_version = 2,
  write_scrap = TRUE,
  low_cell_correction = 0.5,
  low_cell_threshold = 0,
  insensible_rt_lo = 99,
  insensible_rt_hi = 101,
  log_base = exp(1),
  dropRaw = FALSE
)
}
\description{
The \code{mill()} runs the CEDAR pipeline with useful defaults.
}
