% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_blade_depth.R
\name{set_blade_depth}
\alias{set_blade_depth}
\title{Generate a Named Vector Mapping CEDAR Export Column Names to Standardized Timber Column Names}
\usage{
set_blade_depth(cedar_version = 2)
}
\description{
\code{set_blade_depth()} generates a named vector, containing CEDAR export column names as
  values, and their equivalent standardized timber column names as names.

 @param cedar_version
   numeric [1,2]: the version of CEDAR that created the export.

 @details
   The \code{set_blade_depth()} maps original, non-standardized CEDAR export column names to
   standardized timber column names -- for both versions of CEDAR. These column names are provided
   in an external .CSV: \code{/inst/raw_data/import_filter.csv}.

 @return
   A named vector.

 @importFrom readr cols read_csv
 @importFrom tidyr drop_na
 @importFrom dplyr pull %>%

 @export
}
