% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_mill.R
\name{start_mill}
\alias{start_mill}
\title{Start the sawmill Pipeline}
\usage{
start_mill()
}
\description{
The \code{start_mill()} function prompts the user for a CEDAR query, the CEDAR query version,
 and runs the CEDAR pipeline with useful defaults.
}
